integrations/desktop_agent/proxy/requirements.txt
mitmproxy>=10.0.0
psutil>=5.9.0
pyyaml>=6.0
